Inclusao do codigo para chamada da procedure e rotinha para alimentar o grid na tela.

+++ BNP.Teste/BNP.Teste.Service/DTO/ListaMovimentoDto.cs
@@ -2,11 +2,11 @@
 {
     public class ListaMovimentoDto
     {
-        public int Mes { get; set; }
-        public int Ano { get; set; }
+        public decimal Mes { get; set; }
+        public decimal Ano { get; set; }
         public string CodProduto { get; set; }
         public string DescricaoProduto { get; set; }
-        public long NumLancamento { get; set; }
+        public decimal NumLancamento { get; set; }
         public string Descricao { get; set; }
         public decimal Valor { get; set; }
     }
